% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{omicsFixID_app}
\alias{omicsFixID_app}
\title{Launch "omicsFixID" Shiny App}
\usage{
omicsFixID_app(...)
}
\description{
OmicsFixID provides two key functions.
(1)FixID: provides statistics on the percentage of secondary identifiers in the dataset and converts outdated secondary identifiers to current primary identifiers, if available. The FixID functionality currently covers secondary identifiers from HGNC , HMDB , ChEBI and Wikidata which can be converted to the corresponding primary identifier from the initial database. After this step, the CrossMapID can be used to convert the primary-ID-enhanced dataset to any other database currently supported by BridgeDb. (2) CrossMapID: uses BridgeDb's REST-API to convert identifiers.
}
